requirements.txt
Django~=4.2.14
djangorestframework==3.15.2
django-cors-headers==4.4.0

python-dotenv~=1.0.1

celery~=5.4.0
django-celery-beat==2.7.0

requests==2.32.3
psycopg2-binary==2.9.9

gunicorn==22.0.0

drf-yasg==1.21.7  # for swagger
ipython==8.18.1  # for debugging

websockets==15.0.1
redis==5.2.1

channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2